Feat: Add style for header
After adding a "header" class to the H1 element, I changed its color to match the
header design. I noticed that there was a gap between the top/sides of the page and the background
color for the header. To fix this, I adjusted the margins on both the header class
and the body element of the page. Adjusting the margins in both of these places
made the background color of the header flush against the browser window, more similar
to how it appeared in the header design.

I also adjusted the font for the header and used relative sizing to create padding
between the top and the bottom of the header to more closely match the header design.

FSA2021-67