feat: serve 404 not found page for nonexistent endpoint
The current version of the app shows only a message saying "Cannot GET /askldj" if you typed in a route that doesn't exist, for example http://localhost:3000/askldj. This PR displays a more helpful 404 status code and 404 page to tell the user that the resource has not been found.

FSA2021-73