Test: Replace HTML import
As the HTML for our grid is likely to evolve, directly importing the HTML file means that
the HTML our tests are based off of will continuously change. This could cause tests that
pass with consideration to our current code, to fail in the future. To prevent this from
happening, I removed the HTML import that used the file system and replaced it with a
gridGame variable written directly into the describe block for handling the click cell
function.

FSA2021-68