Feat: Add toggling between X and O
Created a state object with a key of currentTurn that will have a value of either X or O.
By default, it has a value of "X." When a cell is clicked, the currentTurn value from this
state object will replace the cell's inner text. A ternary operator updates the state object
for the next turn depending on whether the current turn is an X or an O.

The toggling feature caused tests that were previously written to fail. To correct a test
that used the for each loop attached to allGameCells, I used beforeEach() and afterEach()
to set up and tear down the testing environment. The afterEach() function also uses the
forEach() method to clear each cell and reset the state of the current turn to X, so as to
prevent the next test from returning an "O." Resetting the state required importing the
state object from the JS file. I deleted the final test because its expected behavior
(i.e. clicking on all cells would turn all cells into Xs) was no longer consistent with
the game's logic (i.e. each subsequent click alternating between X and O).

I created additional tests to show toggling behavior, which included creating a new JSDOM using
the current HTML file, and replicating the forEach loop on allGameCells in the JS file.
To test that each subsequent click alternated between X and O, I created a for loop where all
clicks where i % 2 === 0 (i.e. all even clicks, where zero is considered even) should replace the
text content with an X, and all odd clicks should show an O.

After rebasing against the main branch, the linter produced a duplicate object warning on the
package lock file, so I ran `npm i` again to fix it and update the package lock.

FSA2021-69