Refactor: Update constructor function
Removed the argument that constructor function took and added currentPlayer and winner.
As the default value for currentPlayer is 'X,' I added that directly into the constructor
function and removed the new gamestate variable. new Game() will now load with the values
in gamestate by default. I updated everywhere gameState was referenced.