Feat: Add markBoard function
Added a markBoard() function to the game class to mark up the gameBoard array.
To keep tests from failing, I temporarily added in an if/else statement to the
handleClickCell() function that continues to mark up a cell's innerText for all
cells that have an HTML class attached to them. I also reset the current player
for the game to 'X' for the second describe block of tests to keep tests from
failing due to returning the opposite game piece (i.e. 'O' when it should be 'X').