test: set up subheader tests
- Adds some test blocks to state exactly what we intend to test with
this new feature.
- To be able to easily include the multiple apostrophes that are
inevitable when writing these tests, the eslint single quote rule is
disabled for the test file.
- Add (passing) tests to the empty test blocks for the subheader
- Instead of utilizing two JSDOM instances for the subheader and grid,
this combines them into a single JSDOM instance.
- To be able to create a new instance of Game within tests as needed,
exports Game class rather than game const.
- The handleClickCell() function needs to rely on whatever the current
instance of game is. In game.js it defaults to the const game = new
Game() we already created. In index.spec.js it will default to the game
instance we create in the test file.