Refactor: update existing tests
- This builds on a previous refactoring to remove repetitive code by
creating functions for pieces of code that are repeatedly created.
-Condenses separate JSDOM instances being used in tests into one.
- Since subheader tests have been written in a different style to
better utilize beforeEach() (which will minimize data leaking from test
to test), this also updates the style of the tests to match.
- With the second block of tests (Toggling between Xs and Os), this was
not immediately possible without heavier refactoring, so it is left
largely the same. This block would need more refactoring if tests were
to be added to it, but is currently just fine as is.
- As we go forward, it might be helpful to move index.spec.js into a
separate folder for tests. This way, we can run tests specifically for
the Game class in a game.spec.js file and the two files can be together
in one folder. If we have a testing folder, we could potentially have a
separate file for helpful functions for the tests. Here, just
highlighting what these helper functions to clarify, especially in the
event that they are ever separated from index.spec.js.