feat: add Game getters/setters
-In the previous version, currentPlayer and gameBoard are accessed from
handleClickCell(). Here, currentPlayer and gameBoard are renamed with an
underscore in front of them to denote that it is intended to be a
private property.
-In order to access these "private" properties outside of the class,
get currentPlayer(), get gameBoard(), and set currentPlayer() have been
added.
-winner in the constructor has also been changed to _winner to follow
the same convention, but no getters/setters have been added for winner
at this time.

FSA2021-74